antennaCAT compat version update
