+++ README.md
@@ -1,34 +1,84 @@
 # pso_python
-Simple adaptive timestep particle swarm optimizer written in Python
+Simple adaptive timestep particle swarm optimizer written in Python. 
+Now featuring AntennaCAT hooks for GUI integration and user input handling
 
-## Times-step adaptation 
-This particle swarm optimizers using the mean absolute deviation of particle position
-as an adjustment to the time step, to prevent the particle overshoot problem.  This 
-particle distribution is initialized to one when the swarm starts, so that the impact
-is boundary independent. 
+## Table of Contents
+* [Requirements](#requirements)
+* [Implementation](#implementation)
+    * [Time-step adaptation](#time-step-adaptation)
+    * [Constraint Handling](#constraint-handling)
+    * [Boundary Types](#boundary-types)
+    * [Multi-Object Optimization](#multi-object-optimization)
+    * [Objective Function Handling](#objective-function-handling)
+* [Examples](#example-implementations)
+    * [Basic PSO Example](#basic-pso-example)
+    * [Detailed Messages](#detailed-messages)
+    * [Realtime Graph](#realtime-graph)
+* [References](#references)
+* [Publications and Integration](#publications-and-integration)
+* [Licensing](#licensing)  
 
-## Constraint handling
-Users must create thier own constraint function for thier problems, if there are  constraints
-beyond the problem bounds.  This is then passed into the constructor. If the default constraint
-function is used, it always returns true (which means there are no constraints).
+## Requirements
 
-## Boundry types
-This PSO optimizers has 4 different types of bounds, Random (Particles that leave the area respawn),
-Reflection (Particles that hit the bounds reflect), Absorb (Particles that hit the bounds lose velocity
-in that direction), Invisible (Out of bound particles are no longer evaluated).
+This project requires numpy and matplotlib. 
 
-Some updates have not incorporated appropariate handling for all boundry conditions.  This bug is known
-and is being worked on.  The most consistent boundary type at the moment is Random.  If constraints are
-violated, but bounds are not, currently random bound rules are used to deal with this problem. 
+Use 'pip install -r requirements.txt' to install the following dependencies:
 
-## Multi-Object Optimization
-The no preference method of multi-objective optimization, but a Pareto Front is not calculated.
-Instead the best choice (smallest norm of output vectors) is listed as the output.
+```python
+contourpy==1.2.1
+cycler==0.12.1
+fonttools==4.51.0
+importlib_resources==6.4.0
+kiwisolver==1.4.5
+matplotlib==3.8.4
+numpy==1.26.4
+packaging==24.0
+pillow==10.3.0
+pyparsing==3.1.2
+python-dateutil==2.9.0.post0
+six==1.16.0
+zipp==3.18.1
 
-## Objective function handling
-The optimizer minimizes the absolute value of the difference from the targets outputs and the
-evaluated outputs.  Future versions may include options for function minimization absent target
-values. 
+```
 
-## Example implimentation
-PSO_TEST.py provides a sample use case of the optimizer. 
+## Implementation
+### Time-Step Adaptation 
+This particle swarm optimizers using the mean absolute deviation of particle position as an adjustment to the time step, to prevent the particle overshoot problem.  This particle distribution is initialized to one when the swarm starts, so that the impact is boundary independent. 
+
+### Constraint Handling
+Users must create their own constraint function for their problems, if there are constraints beyond the problem bounds.  This is then passed into the constructor. If the default constraint function is used, it always returns true (which means there are no constraints).
+
+### Boundary Types
+This PSO optimizers has 4 different types of bounds, Random (Particles that leave the area respawn), Reflection (Particles that hit the bounds reflect), Absorb (Particles that hit the bounds lose velocity in that direction), Invisible (Out of bound particles are no longer evaluated).
+
+Some updates have not incorporated appropriate handling for all boundary conditions.  This bug is known and is being worked on.  The most consistent boundary type at the moment is Random.  If constraints are violated, but bounds are not, currently random bound rules are used to deal with this problem. 
+
+### Multi-Object Optimization
+The no preference method of multi-objective optimization, but a Pareto Front is not calculated. Instead the best choice (smallest norm of output vectors) is listed as the output.
+
+### Objective Function Handling
+The optimizer minimizes the absolute value of the difference from the target outputs and the evaluated outputs.  Future versions may include options for function minimization absent target values. 
+
+## Example Implementations
+
+### Basic PSO Example
+pso_test.py provides a sample use case of the optimizer. 
+
+### Detailed Messages
+pso_test_details.py provides an example using a parent class, and the self.suppress_output and detailedWarnings flags to control error messages that are passed back to the parent class to be printed with a timestamp. This implementation sets up the hooks for integration with AntennaCAT in order to provide the user feedback of warnings and errors.
+
+### Realtime Graph
+pso_test_graph.py provides an example using a parent class, and the self.suppress_output and detailedWarnings flags to control error messages that are passed back to the parent class to be printed with a timestamp. Additionally, a realtime graph shows particle locations at every step.
+
+NOTE: if you close the graph as the code is running, the code will continue to run, but the graph will not re-open.
+
+
+## References
+TODO: add in references to papers as boundaries and solution options are added.
+
+## Publications and Integration
+This software works as a stand-alone implementation, and as one of the optimizers integrated into AntennaCAT.
+
+Publications featuring the code in this repo will be added as they become public.
+
+## Licensing
