--- constr_default.py
@@ -1,6 +1,0 @@
-#! /usr/bin/python3
-import numpy as np
-
-
-def constr_default(x):
-    return True
